Support liantichess variants without variants.ini configuration.
This is done because fairy-stockfish.wasm does not support configuring variants through the variants.ini file (yet).

Supports all Liantichess variants other than antiempire, antichak and all atomic related variants. This is because the version of fairy stockfish used does not support custom pieces, which is needed for antiempire and antichak, and does not support atomic (blastOnCapture), which is needed for antiatomic, unlike the master version.